+++ packages/app/src/debug/DebugContainer.cy.tsx
@@ -1,10 +1,11 @@
-import { DebugSpecListGroupsFragment, DebugSpecListSpecFragment, DebugSpecListTestsFragment, DebugSpecsFragmentDoc, UseCohorts_DetermineCohortDocument } from '../generated/graphql-test'
+import { DebugSpecListGroupsFragment, DebugSpecListSpecFragment, DebugSpecListTestsFragment, DebugSpecsFragment, DebugSpecsFragmentDoc, UseCohorts_DetermineCohortDocument } from '../generated/graphql-test'
 import DebugContainer from './DebugContainer.vue'
 import { defaultMessages } from '@cy/i18n'
 import { useLoginConnectStore } from '@packages/frontend-shared/src/store/login-connect-store'
 import { specsList } from './utils/DebugMapping'
 import { CloudRunStubs } from '@packages/graphql/test/stubCloudTypes'
 import { DEBUG_SLIDESHOW } from './utils/constants'
+import type { CloudRun, CloudSpecRun, CloudTestResult } from '@packages/graphql/src/gen/test-cloud-graphql-types.gen'
 
 const DebugSpecVariableTypes = {
   hasNextRun: 'Boolean',
@@ -238,7 +239,7 @@ describe('<DebugContainer />', () => {
       loginConnectStore.setHasInitiallyLoaded()
     })
 
-    it('render first pending run', () => {
+    it('renders running run', () => {
       cy.mountFragment(DebugSpecsFragmentDoc, {
         variableTypes: DebugSpecVariableTypes,
         variables: {
@@ -262,14 +263,15 @@ describe('<DebugContainer />', () => {
       })
 
       cy.findByTestId('debug-header').should('be.visible')
+      cy.findByTestId('debug-testing-progress').should('be.visible')
       cy.findByTestId('debug-pending-splash')
       .should('be.visible')
       .within(() => {
-        cy.findByTestId('debug-pending-counts').should('have.text', '0 of 0 specs completed')
+        cy.contains('Testing in progress...')
       })
     })
 
-    it('does not render DebugPendingRunSplash and DebugNewRelevantRunBar at the same time', () => {
+    it('renders running run with failed tests', () => {
       cy.mountFragment(DebugSpecsFragmentDoc, {
         variableTypes: DebugSpecVariableTypes,
         variables: {
@@ -281,23 +283,12 @@ describe('<DebugContainer />', () => {
           if (result.currentProject?.cloudProject?.__typename === 'CloudProject') {
             const test = result.currentProject.cloudProject.runByNumber
 
-            // Testing this to confirm we are "making impossible states impossible" in the UI,
-            // and document the expectation in this scenario. For clarity,
-            // we do not expect a 'RUNNING` current and next run at the same time, so
-            // the data below represents an invalid state.
-
             result.currentProject.cloudProject.runByNumber = {
-              ...CloudRunStubs.running,
+              ...CloudRunStubs.failingWithTests,
+              status: 'RUNNING',
               runNumber: 1,
               completedInstanceCount: 2,
               totalInstanceCount: 3,
-            } as typeof test
-
-            result.currentProject.cloudProject.nextRun = {
-              ...CloudRunStubs.running,
-              runNumber: 1,
-              completedInstanceCount: 5,
-              totalInstanceCount: 6,
             } as typeof test
           }
         },
@@ -305,13 +296,138 @@ describe('<DebugContainer />', () => {
       })
 
       cy.findByTestId('debug-header').should('be.visible')
-      cy.findByTestId('debug-pending-splash')
-      .should('be.visible')
-      .within(() => {
-        cy.findByTestId('debug-pending-counts').should('have.text', '0 of 0 specs completed')
+      cy.findByTestId('debug-testing-progress').should('be.visible')
+      cy.findByTestId('debug-spec-item').should('be.visible')
+    })
+
+    it('simulates full running run with failed tests', () => {
+      //Change this value when modifying this test to allow time to see transitions
+      const waitTimeBetweenSimulatedEvents = 0
+
+      cy.mountFragment(DebugSpecsFragmentDoc, {
+        variableTypes: DebugSpecVariableTypes,
+        variables: {
+          hasNextRun: false,
+          runNumber: 1,
+          nextRunNumber: -1,
+        },
+        onResult: (result) => {
+          if (result.currentProject?.cloudProject?.__typename === 'CloudProject') {
+            const test = result.currentProject.cloudProject.runByNumber
+
+            //creating copy to prevent mutation later on in this test
+            const cloudRunCopy = JSON.parse(JSON.stringify(CloudRunStubs.failingWithTests))
+
+            result.currentProject.cloudProject.runByNumber = {
+              ...cloudRunCopy,
+              status: 'RUNNING',
+              runNumber: 1,
+              completedInstanceCount: 2,
+              totalInstanceCount: 3,
+            } as typeof test
+          }
+        },
+        render: (gqlVal) => {
+          return <DebugContainer gql={gqlVal} />
+        },
+      }).then(({ component }) => {
+        cy.wrap(component.gql).as('gql')
       })
 
-      cy.findByTestId('newer-relevant-run').should('not.exist')
+      /*
+        DebugTestingProgress is not mocked here resulting in it always
+        saying "0 of 0 specs completed"
+        Trying to mock up the DebugTestingProgress here results in
+        Urql clearing out and refetching the main query which makes
+        this test fail.
+      */
+
+      const getRun = (gql: DebugSpecsFragment) => {
+        const run = gql.currentProject?.cloudProject?.__typename === 'CloudProject'
+        && gql.currentProject.cloudProject.runByNumber
+
+        if (!run) {
+          throw Error('Could not find run')
+        }
+
+        return run
+      }
+
+      cy.findByTestId('debug-header').should('be.visible')
+      cy.findByTestId('debug-testing-progress').should('be.visible')
+
+      cy.wait(waitTimeBetweenSimulatedEvents)
+      cy.get<DebugSpecsFragment>('@gql').then((gql) => {
+        const run = getRun(gql)
+
+        cy.wrap(run).as('run')
+
+        run.totalFailed = 4
+        const newTest = JSON.parse(JSON.stringify(run.testsForReview[0]))
+
+        newTest.id = '345'
+        newTest.thumbprint = `${newTest.thumbprint}c`
+
+        run.testsForReview.push(newTest)
+      })
+
+      cy.wait(waitTimeBetweenSimulatedEvents)
+      cy.findAllByTestId('debug-spec-item').should('have.length', 1)
+      cy.get<CloudRun>('@run').then((run) => {
+        const newSpec: CloudSpecRun = JSON.parse(JSON.stringify(run.specs[0]))
+
+        newSpec.id = 'spec2'
+        run.specs.push(newSpec)
+        cy.wrap(newSpec).as('newSpec')
+
+        const newSpecTest: CloudTestResult = JSON.parse(JSON.stringify(run.testsForReview[0]))
+
+        newSpecTest.id = '789'
+        newSpecTest.thumbprint = `${newSpecTest.thumbprint}d`
+        newSpecTest.specId = newSpec.id
+        run.testsForReview.push(newSpecTest)
+
+        cy.wrap(newSpecTest).as('newSpecTest')
+      })
+
+      cy.wait(waitTimeBetweenSimulatedEvents)
+      cy.findAllByTestId('debug-spec-item').should('have.length', 2)
+      cy.get<CloudRun>('@run').then((run) => {
+        const newGroup = JSON.parse(JSON.stringify(run.groups[0]))
+
+        newGroup.id = 'Group2'
+        newGroup.groupName = 'Group-Linux-Electron'
+        newGroup.os.name = 'Linux'
+        newGroup.os.platform = 'LINUX'
+        newGroup.os.version = '16'
+        newGroup.os.nameWithVersion = 'Linux 16'
+        run.groups.push(newGroup)
+
+        cy.get<CloudSpecRun>('@newSpec').then((newSpec) => {
+          newSpec.groupIds!.push(newGroup.id)
+
+          cy.get<CloudTestResult>('@newSpecTest').then((newSpecTest) => {
+            const newGroupTest = JSON.parse(JSON.stringify(newSpecTest))
+
+            newGroupTest.instance.groupId = newGroup.id
+
+            run.testsForReview.push(newGroupTest)
+          })
+        })
+      })
+
+      cy.findAllByTestId('debug-spec-item').should('have.length', 2)
+      cy.get(':nth-child(2) > [data-cy="debug-spec-item"] > [data-cy="test-group"] > [data-cy="debug-failed-test-groups"]')
+      .within(() => {
+        cy.findAllByTestId('grouped-row').should('have.length', 2)
+      })
+
+      cy.wait(waitTimeBetweenSimulatedEvents)
+      cy.get<CloudRun>('@run').then((run) => {
+        run.status = 'FAILED'
+      })
+
+      cy.findByTestId('debug-testing-progress').should('not.exist')
     })
 
     it('renders specs and tests when completed run available', () => {
@@ -423,7 +539,7 @@ describe('<DebugContainer />', () => {
         cy.findByTestId('newer-relevant-run')
         .should('be.visible')
         .and('contain.text', 'fix: make gql work FAILED')
-        .and('contain.text', 'View run')
+        .and('contain.text', 'Switch to run')
       })
     })
   })