+++ packages/app/src/debug/LayeredBrowserIcons.cy.tsx
@@ -2,13 +2,13 @@ import LayeredBrowserIcons from './LayeredBrowserIcons.vue'
 import type { BrowserType } from './LayeredBrowserIcons.vue'
 
 describe('<LayeredBrowserIcons/>', () => {
-  const browsers: BrowserType[] = ['CHROME', 'CHROME-CANARY', 'FIREFOX', 'WEBKIT', 'EDGE', 'ELECTRON']
+  const browsers: BrowserType[] = ['CHROME', 'CHROME-CANARY', 'CHROME BETA', 'FIREFOX', 'WEBKIT', 'EDGE', 'ELECTRON']
 
   it('mounts correctly for single browser', () => {
     browsers.forEach((ele) => {
       cy.mount(() => (
         <div class='bg-gray-50 p-24px'>
-          <LayeredBrowserIcons order={[ele]}/>
+          <LayeredBrowserIcons browsers={[ele]}/>
         </div>
       ))
 
@@ -19,12 +19,12 @@ describe('<LayeredBrowserIcons/>', () => {
   it('mounts correctly for multiple browsers', () => {
     cy.mount(() => (
       <div class='bg-gray-50 p-24px'>
-        <LayeredBrowserIcons order={browsers} />
-        <LayeredBrowserIcons order={['CHROME', 'FIREFOX', 'EDGE']}/>
-        <LayeredBrowserIcons order={['CHROME-CANARY', 'WEBKIT', 'ELECTRON']} />
-        <LayeredBrowserIcons order={['FIREFOX', 'WEBKIT', 'EDGE']}/>
-        <LayeredBrowserIcons order={['FIREFOX', 'EDGE', 'CHROME-CANARY', 'WEBKIT']}/>
-        <LayeredBrowserIcons order={['EDGE', 'CHROME', 'WEBKIT', 'ELECTRON']}/>
+        <LayeredBrowserIcons browsers={browsers} />
+        <LayeredBrowserIcons browsers={['CHROME', 'FIREFOX', 'EDGE']}/>
+        <LayeredBrowserIcons browsers={['CHROME-CANARY', 'WEBKIT', 'ELECTRON']} />
+        <LayeredBrowserIcons browsers={['FIREFOX', 'WEBKIT', 'EDGE']}/>
+        <LayeredBrowserIcons browsers={['FIREFOX', 'EDGE', 'CHROME-CANARY', 'WEBKIT']}/>
+        <LayeredBrowserIcons browsers={['EDGE', 'CHROME', 'WEBKIT', 'ELECTRON']}/>
       </div>
     ))